added check for existing directory

+++ cmg.js
@@ -13,6 +13,13 @@ const project_name = process.argv[2];
 
 const current_dir = process.cwd();
 const project_dir = path.resolve(current_dir, project_name);
+// check if directory alread exist
+if (fs.existsSync(project_dir))
+{
+    console.log(`directory ${project_name} already exist, error`);
+    process.exit(1);
+}
+
 fs.mkdirSync(project_dir, { recursive: true });
 const template_dir = path.resolve(__dirname, 'template');
 fs.cpSync(template_dir, project_dir, { recursive: true });
@@ -31,3 +38,5 @@ fs.writeFileSync(
 
 console.log('Success! Your new Base Phaser 3 Game is ready.');
 console.log(`Created ${project_name} at ${project_dir}`);
+console.log(`Proceded with cd ${project_name}`);
+console.log(`npm install`);
